+++ spec/dummy/app/models/comment_reply.rb
@@ -1,0 +1,4 @@
+class CommentReply < ApplicationRecord
+  belongs_to :parent, class_name: 'Comment'
+  belongs_to :child, class_name: 'Comment'
+end