+++ spec/dummy/app/serializers/api/v1/comment_reply_serializer.rb
@@ -1,0 +1,8 @@
+module API
+  module V1
+    class CommentReplySerializer < ApplicationSerializer
+      belongs_to :parent, serializer: CommentSerializer
+      belongs_to :child, serializer: CommentSerializer
+    end
+  end
+end