+++ spec/dummy/app/serializers/api/v1/comment_serializer.rb
@@ -7,6 +7,7 @@ module API
       belongs_to :user
 
       has_one :rating
+      has_one :comment_reply
     end
   end
 end