+++ spec/dummy/db/migrate/20200704003218_create_comment_replies.rb
@@ -1,0 +1,8 @@
+class CreateCommentReplies < ActiveRecord::Migration[5.1]
+  def change
+    create_table :comment_replies do |t|
+      t.integer :parent_id
+      t.integer :child_id
+    end
+  end
+end