+++ spec/dummy/db/schema.rb
@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,40 +10,45 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170125204237) do
+ActiveRecord::Schema.define(version: 20200704003218) do
 
   create_table "attachments", force: :cascade do |t|
-    t.string   "type"
-    t.string   "name"
-    t.integer  "post_id"
+    t.string "type"
+    t.string "name"
+    t.integer "post_id"
     t.datetime "created_at"
     t.datetime "updated_at"
+  end
+
+  create_table "comment_replies", force: :cascade do |t|
+    t.integer "parent_id"
+    t.integer "child_id"
   end
 
   create_table "comments", force: :cascade do |t|
-    t.string   "body",       null: false
+    t.string "body", null: false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "post_id"
-    t.integer  "user_id"
+    t.integer "post_id"
+    t.integer "user_id"
   end
 
   create_table "posts", force: :cascade do |t|
-    t.string   "title",      null: false
+    t.string "title", null: false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "user_id"
+    t.integer "user_id"
   end
 
   create_table "ratings", force: :cascade do |t|
-    t.integer  "value",      null: false
+    t.integer "value", null: false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "comment_id"
+    t.integer "comment_id"
   end
 
   create_table "users", force: :cascade do |t|
-    t.string   "name",       null: false
+    t.string "name", null: false
     t.datetime "created_at"
     t.datetime "updated_at"
   end