+++ spec/support/factories/post.rb
@@ -13,6 +13,16 @@ FactoryGirl.define do
       end
     end
 
+    trait :with_comments_and_replies do
+      transient do
+        comment_count 2
+      end
+
+      after(:create) do |post, evaluator|
+        create_list(:comment, evaluator.comment_count, :with_replies, post: post)
+      end
+    end
+
     trait :with_attachments do
       transient do
         attachment_count 1