+++ CHANGELOG.md
@@ -196,3 +196,4 @@
   * Old functionality: Called after `assign_changes_from_document`, same functionality as `before_create`
 * Add `relationships_referenced` helper so that when `optimize_relationships` is enabled and relationships are changed as part of a persistence request (create, update), respond with the data for these relationships in the primary data instead of skipping them
 * BREAKING: Use ActionController::API for Caprese::Controller
+* Allow includes to overwrite past includes of same resource identifier